fix: major overhaul
